Fix render issue when button has conditional
If the button has a conditional, the image render function will
return a 0 which indicated that an image was present even if it
was not if the condition was not being met during boot. When the
condition is met, mButtonImg was not set to NULL causing a render
request failure.

Change-Id: I32be0ddcdc1bf7ad22440d66018ef09484280afa