Revert "Force sync files written by minzip."
This has been causing significant performance degradation when flashing updates
for the entirety of 5.0.

I was originally hesitant to do this since we don't want unsynced files on an update,
however - Our updater-script unmounts /system at the end of installation, which
will force a sync.  As a result the issue Google was trying to fix won't be present
for our ZIPs.

This reverts commit a6c142f2a579ea5e7cdfbc88e6a061c55029265a.

Change-Id: I89b22e2677c4eda4d3e6338adc0046bf754d43b2