Remove outdated reference to icon_installing.png
The Wear recovery UI doesn't draw the installing icon but it was still
trying to open it.  Ever since these images were removed eight months
ago, this has resulted in an error printing to the screen at runtime.
Since the image wasn't really used, the lines to open it can simply be
removed.

Bug: 33203397
Change-Id: Id820f6d75e316c51d19b6095df407ecd61c0410e