recovery: secure MTP and ADB when pass/pattern set
ADB
-----
1. on boot SHRP will check if pass/pattern is set
2. depending on 1. the property "shrp.lock" will be set
   to 0 (off) or 1 (on)
3. ADB will not start when a pass/pattern is set
4. when the correct pass/pattern is entered shrp.lock will
   change to 0
5. and finally on property shrp.lock=0 ADB will start

MTP
-----
1. same as for ADB
2. depending on 1. "lock_enabled" is set to 0 (off) or 1 (on)
3. MTP will not start when a pass/pattern is set (boot)
4. MTP will not start until the correct pass/pattern has been
   entered (handled directly within startmtp function)
5. triggered by ADB MTP will be started*

for 5. a better implementation would be to move the whole
"#ifdef TW_HAS_MTP" part of main() in twrp.cpp to a new function
and calling that function there AND within lockCheck() to
handle non-MTP devices properly.

note:
the whole "security" setting within a recovery is just
useful for one specific use-case:

*protecting your phone when the attacker do not have access
to it by fastboot*

as soon as the attacker has access by fastboot any security
setting is useless as you can overwrite the recovery by whatever
you like. that's a matter of fact which simply comes with
unlocking the bootloader.

anyways this implementation ensures that an attacker NEEDS to
replace the recovery in order to access your files and so makes
it a little bit harder for being hacked.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>